add jwt

+++ main.go
@@ -10,15 +10,16 @@ import (
 	. "github.com/Monibuca/utils/v3"
 	"github.com/tidwall/gjson"
 	"github.com/tidwall/sjson"
+	"github.com/yunnet/plugin-firefly/jwt"
 	result "github.com/yunnet/plugin-firefly/web"
 	"io"
-	"strings"
-
 	"io/ioutil"
+	"log"
 	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"strings"
 )
 
 var config struct {
@@ -27,11 +28,12 @@ var config struct {
 	Password string
 }
 
-const C_JSON_FILE = "firefly.json"
-const C_NETWORK_HEAD = "iface eth0"
-const C_NETWORK_FILE = "/etc/network/interfaces"
-
-const C_SALT = "firefly"
+const (
+	C_JSON_FILE    = "firefly.json"
+	C_NETWORK_HEAD = "iface eth0"
+	C_NETWORK_FILE = "/etc/network/interfaces"
+	C_SALT         = "firefly"
+)
 
 func init() {
 	InstallPlugin(&PluginConfig{
@@ -49,25 +51,83 @@ func run() {
 	//登陆
 	http.HandleFunc("/api/firefly/login", getLoginHandler)
 
-	//网络查询
-	http.HandleFunc("/api/firefly/config/tcp", getConfigTcpHandler)
-	//网络设置
-	http.HandleFunc("/api/firefly/config/tcp/edit", editConfigTcpHandler)
-
-
 	//JSON配置查询
 	http.HandleFunc("/api/firefly/config", getConfigHandler)
 	//JSON配置编辑
 	http.HandleFunc("/api/firefly/config/edit", editConfigHandler)
 
+	//网络查询
+	http.HandleFunc("/api/firefly/config/tcp", getConfigTcpHandler)
+	//网络设置
+	http.HandleFunc("/api/firefly/config/tcp/edit", editConfigTcpHandler)
+
+	//http.HandleFunc("/api/firefly/config/ping", pingConfigTcpHandler)
+}
+
+//func pingConfigTcpHandler(w http.ResponseWriter, r *http.Request) {
+//
+//}
+
+func checkLogin(w http.ResponseWriter, r *http.Request) bool {
+	tokenString := r.Header.Get("token")
+
+	valid, err := jwt.ValidateToken(tokenString)
+	if err != nil {
+		res := result.ErrUnauthorized
+		w.Write(res.Raw())
+
+		log.Println(err.Error())
+		return false
+	}
+	return valid
+}
+
+func getLoginHandler(w http.ResponseWriter, r *http.Request) {
+	CORS(w, r)
+	requestUser := r.URL.Query().Get("username")
+	if requestUser == "" {
+		res := result.Err.WithMsg("用户名不能为空")
+		w.Write(res.Raw())
+		return
+	}
+	requestPassword := r.URL.Query().Get("password")
+	if requestPassword == "" {
+		res := result.Err.WithMsg("密码不能为空")
+		w.Write(res.Raw())
+		return
+	}
+
+	user := config.Username
+	if user != requestUser {
+		res := result.Err.WithMsg("用户名或密码错误,请重新输入")
+		w.Write(res.Raw())
+		return
+	}
+
+	m5 := md5.New()
+	m5.Write([]byte(requestPassword + C_SALT))
+	password := hex.EncodeToString(m5.Sum(nil))
+	if config.Password != password {
+		res := result.Err.WithMsg("用户名或密码错误,请重新输入")
+		w.Write(res.Raw())
+		return
+	}
+	tokenString, _ := jwt.CreateToken(user, 3600)
+
+	res := result.OK.WithData(tokenString)
+	w.Write(res.Raw())
 }
 
 func rebootHandler(w http.ResponseWriter, r *http.Request) {
 	CORS(w, r)
+	isOk := checkLogin(w, r)
+	if !isOk {
+		return
+	}
 
 	cmd := exec.Command("reboot")
 	console, err := cmd.Output()
-	if err != nil{
+	if err != nil {
 		res := result.Err.WithMsg(err.Error())
 		w.Write(res.Raw())
 		return
@@ -77,8 +137,178 @@ func rebootHandler(w http.ResponseWriter, r *http.Request) {
 	w.Write(res.Raw())
 }
 
+func getConfigHandler(w http.ResponseWriter, r *http.Request) {
+	CORS(w, r)
+	isOk := checkLogin(w, r)
+	if !isOk {
+		return
+	}
+
+	if node := r.URL.Query().Get("node"); node != "" {
+		filePath := filepath.Join(config.Path, C_JSON_FILE)
+		content, err := readFile(filePath)
+		if nil != err {
+			res := result.Err.WithMsg(err.Error())
+			w.Write(res.Raw())
+			return
+		}
+		root := gjson.Parse(content)
+		nodeJson := root.Get(node)
+		res := result.OK.WithData(nodeJson.Value())
+		w.Write(res.Raw())
+	} else {
+		res := result.Err.WithMsg("no such node")
+		w.Write(res.Raw())
+	}
+}
+
+func readFile(filePath string) (content string, err error) {
+	res, err := ioutil.ReadFile(filePath)
+	if nil != err {
+		return "", err
+	}
+	return string(res), nil
+}
+
+func editConfigHandler(w http.ResponseWriter, r *http.Request) {
+	CORS(w, r)
+	isOk := checkLogin(w, r)
+	if !isOk {
+		return
+	}
+
+	if err := r.ParseForm(); err != nil {
+		res := result.Err.WithMsg(err.Error())
+		w.Write(res.Raw())
+		return
+	}
+
+	defer r.Body.Close()
+
+	request, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		res := result.Err.WithMsg(err.Error())
+		w.Write(res.Raw())
+		return
+	}
+	requestJson := gjson.Parse(string(request))
+
+	nodePath := requestJson.Get("node")
+	if !nodePath.Exists() {
+		res := result.Err.WithMsg("param node does not exist.")
+		w.Write(res.Raw())
+		return
+	}
+	nodeData := requestJson.Get("data")
+	if !nodePath.Exists() {
+		res := result.Err.WithMsg("param node data does not exist.")
+		w.Write(res.Raw())
+		return
+	}
+
+	filePath := filepath.Join(config.Path, C_JSON_FILE)
+	content, err := readFile(filePath)
+
+	rootJson := gjson.Parse(content)
+	node := rootJson.Get(nodePath.String())
+	if !node.Exists() {
+		res := result.Err.WithMsg("node does not exist.")
+		w.Write(res.Raw())
+		return
+	}
+
+	resultJson, _ := sjson.Set(content, nodePath.String(), nodeData.Value())
+
+	flag := os.O_CREATE | os.O_TRUNC | os.O_WRONLY
+	file, err := os.OpenFile(filePath, flag, 0755)
+	if err != nil {
+		res := result.Err.WithMsg(err.Error())
+		w.Write(res.Raw())
+		return
+	}
+
+	file.Write([]byte(resultJson))
+	file.Close()
+
+	res := result.OK.WithData("success")
+	w.Write(res.Raw())
+}
+
+func getConfigTcpHandler(w http.ResponseWriter, r *http.Request) {
+	CORS(w, r)
+	isOk := checkLogin(w, r)
+	if !isOk {
+		return
+	}
+
+	file := C_NETWORK_FILE
+	fmt.Println("read file:" + file)
+
+	content, err := readInterfaces(file)
+	if err != nil {
+		res := result.Err.WithMsg(err.Error())
+		w.Write(res.Raw())
+		return
+	}
+	rootJson := gjson.Parse(content)
+	res := result.OK.WithData(rootJson.Value())
+	w.Write(res.Raw())
+}
+
+func readInterfaces(filePath string) (string, error) {
+	f, err := os.Open(filePath)
+	if err != nil {
+		return "", err
+	}
+	defer f.Close()
+
+	s := bufio.NewScanner(f)
+	ready := false
+
+	var ipAddr = make(map[string]interface{}, 4)
+	for s.Scan() {
+		line := s.Text()
+		if strings.HasPrefix(strings.TrimSpace(line), "#") {
+			continue
+		}
+		// Continue if line is empty
+		if len(strings.TrimSpace(line)) == 0 {
+			if ready {
+				break
+			}
+			continue
+		}
+
+		if strings.Contains(line, C_NETWORK_HEAD) {
+			ready = true
+		}
+
+		if ready {
+			sline := strings.Split(strings.TrimSpace(line), " ")
+			switch sline[0] {
+			case "address":
+				ipAddr["address"] = sline[1]
+			case "netmask":
+				ipAddr["netmask"] = sline[1]
+			case "gateway":
+				ipAddr["gateway"] = sline[1]
+			case "dns-nameservers":
+				ipAddr["dns-nameservers"] = sline[1]
+			default:
+			}
+		}
+	}
+	rootJson, _ := json.Marshal(ipAddr)
+	return string(rootJson), nil
+}
+
 func editConfigTcpHandler(w http.ResponseWriter, r *http.Request) {
 	CORS(w, r)
+	isOk := checkLogin(w, r)
+	if !isOk {
+		return
+	}
+
 	if err := r.ParseForm(); err != nil {
 		res := result.Err.WithMsg(err.Error())
 		w.Write(res.Raw())
@@ -107,7 +337,6 @@ func editConfigTcpHandler(w http.ResponseWriter, r *http.Request) {
 	}
 	defer in.Close()
 
-	//flag := os.O_CREATE | os.O_TRUNC | os.O_WRONLY
 	flag := os.O_RDWR | os.O_CREATE
 	out, err := os.OpenFile(fileName, flag, 0755)
 	if err != nil {
@@ -199,189 +428,4 @@ func editConfigTcpHandler(w http.ResponseWriter, r *http.Request) {
 
 	res := result.OK.WithMsg("修改成功")
 	w.Write(res.Raw())
-}
-
-func getLoginHandler(w http.ResponseWriter, r *http.Request) {
-	CORS(w, r)
-	requestUser := r.URL.Query().Get("username")
-	if requestUser == "" {
-		res := result.Err.WithMsg("用户名不能为空")
-		w.Write(res.Raw())
-		return
-	}
-	requestPassword := r.URL.Query().Get("password")
-	if requestPassword == "" {
-		res := result.Err.WithMsg("密码不能为空")
-		w.Write(res.Raw())
-		return
-	}
-
-	user := config.Username
-	if user != requestUser {
-		res := result.Err.WithMsg("用户名或密码错误,请重新输入")
-		w.Write(res.Raw())
-		return
-	}
-
-	m5 := md5.New()
-	m5.Write([]byte(requestPassword + C_SALT))
-	password := hex.EncodeToString(m5.Sum(nil))
-	if config.Password != password {
-		res := result.Err.WithMsg("用户名或密码错误,请重新输入")
-		w.Write(res.Raw())
-		return
-	}
-
-	res := result.OK.WithMsg("登陆成功")
-	w.Write(res.Raw())
-}
-
-func getConfigHandler(w http.ResponseWriter, r *http.Request) {
-	CORS(w, r)
-	if node := r.URL.Query().Get("node"); node != "" {
-		filePath := filepath.Join(config.Path, C_JSON_FILE)
-		content, err := readFile(filePath)
-		if nil != err {
-			res := result.Err.WithMsg(err.Error())
-			w.Write(res.Raw())
-			return
-		}
-		root := gjson.Parse(content)
-		nodeJson := root.Get(node)
-		res := result.OK.WithData(nodeJson.Value())
-		w.Write(res.Raw())
-	} else {
-		res := result.Err.WithMsg("no such node")
-		w.Write(res.Raw())
-	}
-}
-
-func readFile(filePath string) (content string, err error) {
-	res, err := ioutil.ReadFile(filePath)
-	if nil != err {
-		return "", err
-	}
-	return string(res), nil
-}
-
-func editConfigHandler(w http.ResponseWriter, r *http.Request) {
-	if err := r.ParseForm(); err != nil {
-		res := result.Err.WithMsg(err.Error())
-		w.Write(res.Raw())
-		return
-	}
-
-	defer r.Body.Close()
-
-	CORS(w, r)
-	request, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		res := result.Err.WithMsg(err.Error())
-		w.Write(res.Raw())
-		return
-	}
-	requestJson := gjson.Parse(string(request))
-
-	nodePath := requestJson.Get("node")
-	if !nodePath.Exists() {
-		res := result.Err.WithMsg("param node not exists")
-		w.Write(res.Raw())
-		return
-	}
-	nodeData := requestJson.Get("data")
-	if !nodePath.Exists() {
-		res := result.Err.WithMsg("param node data not exists")
-		w.Write(res.Raw())
-		return
-	}
-
-	filePath := filepath.Join(config.Path, C_JSON_FILE)
-	content, err := readFile(filePath)
-
-	rootJson := gjson.Parse(content)
-	node := rootJson.Get(nodePath.String())
-	if !node.Exists() {
-		res := result.Err.WithMsg("find node not exists")
-		w.Write(res.Raw())
-		return
-	}
-
-	resultJson, _ := sjson.Set(content, nodePath.String(), nodeData.Value())
-
-	flag := os.O_CREATE | os.O_TRUNC | os.O_WRONLY
-	file, err := os.OpenFile(filePath, flag, 0755)
-	if err != nil {
-		res := result.Err.WithMsg(err.Error())
-		w.Write(res.Raw())
-		return
-	}
-
-	file.Write([]byte(resultJson))
-	file.Close()
-
-	res := result.OK.WithData("success")
-	w.Write(res.Raw())
-}
-
-func getConfigTcpHandler(w http.ResponseWriter, r *http.Request) {
-	CORS(w, r)
-	file := C_NETWORK_FILE
-	fmt.Println("read file:" + file)
-
-	content, err := readInterfaces(file)
-	if err != nil{
-		res := result.Err.WithMsg(err.Error())
-		w.Write(res.Raw())
-		return
-	}
-	rootJson := gjson.Parse(content)
-	res := result.OK.WithData(rootJson.Value())
-	w.Write(res.Raw())
-}
-
-func readInterfaces(filePath string) (string, error) {
-	f, err := os.Open(filePath)
-	if err != nil {
-		return "", err
-	}
-	defer f.Close()
-
-	s := bufio.NewScanner(f)
-	ready := false
-
-	var ipAddr = make(map[string]interface{}, 4)
-	for s.Scan() {
-		line := s.Text()
-		if strings.HasPrefix(strings.TrimSpace(line), "#") {
-			continue
-		}
-		// Continue if line is empty
-		if len(strings.TrimSpace(line)) == 0 {
-			if ready {
-				break
-			}
-			continue
-		}
-
-		if strings.Contains(line, C_NETWORK_HEAD) {
-			ready = true
-		}
-
-		if ready {
-			sline := strings.Split(strings.TrimSpace(line), " ")
-			switch sline[0] {
-			case "address":
-				ipAddr["address"] = sline[1]
-			case "netmask":
-				ipAddr["netmask"] = sline[1]
-			case "gateway":
-				ipAddr["gateway"] = sline[1]
-			case "dns-nameservers":
-				ipAddr["dns-nameservers"] = sline[1]
-			default:
-			}
-		}
-	}
-	rootJson, _ := json.Marshal(ipAddr)
-	return string(rootJson), nil
 }
